ref(snuba-admin): Edit predefined queries
