+++ snuba/admin/clickhouse/system_queries.py
@@ -1,10 +1,9 @@
 import re
-from dataclasses import dataclass
-from typing import Dict, Optional, Sequence, Type
 
 from clickhouse_driver.errors import ErrorCodes
 
 from snuba.admin.clickhouse.common import InvalidCustomQuery, get_ro_node_connection
+from snuba.admin.clickhouse.predefined_system_queries import SystemQuery
 from snuba.clickhouse.errors import ClickhouseError
 from snuba.clickhouse.native import ClickhouseResult
 from snuba.utils.serializable_exception import SerializableException
@@ -16,74 +15,6 @@ class NonExistentSystemQuery(SerializableException):
 
 class InvalidResultError(SerializableException):
     pass
-
-
-class _QueryRegistry:
-    """Keep a mapping of SystemQueries to their names"""
-
-    def __init__(self) -> None:
-        self.__mapping: Dict[str, Type["SystemQuery"]] = {}
-
-    def register_class(self, cls: Type["SystemQuery"]) -> None:
-        existing_class = self.__mapping.get(cls.__name__)
-        if not existing_class:
-            self.__mapping[cls.__name__] = cls
-
-    def get_class_by_name(self, cls_name: str) -> Optional[Type["SystemQuery"]]:
-        return self.__mapping.get(cls_name)
-
-    @property
-    def all_queries(self) -> Sequence[Type["SystemQuery"]]:
-        return list(self.__mapping.values())
-
-
-_QUERY_REGISTRY = _QueryRegistry()
-
-
-@dataclass
-class SystemQuery:
-    sql: str
-
-    @classmethod
-    def to_json(cls) -> Dict[str, Optional[str]]:
-        return {
-            "sql": cls.sql,
-            "description": cls.__doc__,
-            "name": cls.__name__,
-        }
-
-    def __init_subclass__(cls) -> None:
-        _QUERY_REGISTRY.register_class(cls)
-        return super().__init_subclass__()
-
-    @classmethod
-    def from_name(cls, name: str) -> Optional[Type["SystemQuery"]]:
-        return _QUERY_REGISTRY.get_class_by_name(name)
-
-    @classmethod
-    def all_queries(cls) -> Sequence[Type["SystemQuery"]]:
-        return _QUERY_REGISTRY.all_queries
-
-
-class CurrentMerges(SystemQuery):
-    """Currently executing merges"""
-
-    sql = """
-        SELECT
-          count(),
-          is_currently_executing
-        FROM system.replication_queue
-        GROUP BY is_currently_executing
-        """
-
-
-class ActivePartitions(SystemQuery):
-    sql = """ SELECT
-        active,
-        count()
-    FROM system.parts
-    GROUP BY active
-    """
 
 
 def _run_sql_query_on_host(
@@ -124,7 +55,7 @@ SYSTEM_QUERY_RE = re.compile(
         (FROM|from)
         \s
         system.[a-z_]+
-        (?P<extra>\s[\w\s,=()*+<>'%\-\/]+)?
+        (?P<extra>\s[\w\s,=()*+<>'%"\-\/]+)?
         ;? # Optional semicolon
         $ # End
     """,
@@ -144,7 +75,7 @@ def run_system_query_on_host_with_sql(
         # Don't send error to Snuba if it is an unknown table or column as it
         # will be too noisy
         if exc.code in (ErrorCodes.UNKNOWN_TABLE, ErrorCodes.UNKNOWN_IDENTIFIER):
-            raise InvalidCustomQuery("Invalid query: {exc.message} {exc.code}")
+            raise InvalidCustomQuery(f"Invalid query: {exc.message} {exc.code}")
 
         raise
